Finished tests and functionality.

+++ src/main/java/controllers/AccountController.java
@@ -1,18 +1,23 @@
 package controllers;
 
 import dao.AccountDao;
+import dao.ClientDao;
 import io.javalin.http.Context;
 import models.Account;
 import models.Category;
 import models.Client;
+import org.apache.log4j.Logger;
 import service.AccountService;
+import service.ClientService;
 
 import java.util.List;
 import java.util.Locale;
 
 //connects to dispatcher
 public class AccountController {
-    public static AccountService accountService = new AccountService(new AccountDao());
+    public static AccountService accountService = new AccountService(new AccountDao(), new ClientDao());
+    public static ClientService clientService = new ClientService(new ClientDao(), new AccountDao());
+    public static Logger log = Logger.getLogger(ClientDao.class);
 
     //get all accounts belonging to client
     public static void getAccounts(Context context) {
@@ -28,9 +33,10 @@ public class AccountController {
             lowerLimit = -1.0;
         }
         List<Account> accountList = accountService.getAccounts(clientId, upperLimit, lowerLimit);
-        if(accountList.size() == 0)
+        if(accountList == null)
         {
             context.status(404);
+            context.result("Account not found.");
         }else
         {
             context.contentType("application/json");
@@ -51,6 +57,10 @@ public class AccountController {
             context.contentType("application/json");
             context.status(201);
             context.result(ClientController.convertToJson(successful));
+        }else{
+            context.contentType("application/json");
+            context.status(404);
+            context.result(ClientController.convertToJson(successful));
         }
     }
 
@@ -62,27 +72,49 @@ public class AccountController {
         newAccount.setId(id);
         newAccount.setClientId(clientId);
 
+        Client client = clientService.getClient(clientId);
         Account account = accountService.getAccount(id);
-        if(account == null){
+        if(account == null || client == null){
             context.status(404);
-        }else {
+            context.result("Client or account are not found.");
+            log.error("Account or client not found.");
+
+        }else if (account.getClientId() != clientId){
+            context.status(404);
+            context.result("Account/Client combination not found.");
+        }else{
             context.contentType("application/json");
             context.result(ClientController.convertToJson(account));
         }
     }
 
     public static void updateAccount(Context context) {
-        Integer clientId = Integer.parseInt(context.pathParam("id"));
-        Integer accountId = Integer.parseInt(context.pathParam("accountId"));
-        String name = context.formParam("name");
-        Double balance = Double.parseDouble(context.formParam("balance"));
-        Category category = Category.valueOf(context.formParam("category").toUpperCase());
-        Account account = new Account(name, balance, category, clientId);
-        account.setId(accountId);
+        Integer clientId = 0;
+        Integer accountId = 0;
+        String name = "";
+        Double balance = 0.0;
+        Category category = null;
+        Account account = null;
+        Account updatedAccount = null;
+        try {
+            clientId = Integer.parseInt(context.pathParam("id"));
+            accountId = Integer.parseInt(context.pathParam("accountId"));
+            name = context.formParam("name");
+            balance = Double.parseDouble(context.formParam("balance"));
+            category = Category.valueOf(context.formParam("category").toUpperCase());
+            account = new Account(accountId, name, balance, category, clientId);
 
-        Account updatedAccount = accountService.updateAccount(account);
+            updatedAccount = accountService.updateAccount(account);
+        }catch(Exception e)
+        {
+            //e.printStackTrace();
+            context.result("Error with input.  Check form or url for invalid or missing input.");
+            return;
+        }
+
         if(updatedAccount == null){
             context.status(404);
+            context.result("Account or Client not found.");
         }else {
             context.contentType("application/json");
             context.result(ClientController.convertToJson(updatedAccount));
@@ -91,13 +123,19 @@ public class AccountController {
 
     public static void deleteAccount(Context context) {
         Integer accountId = Integer.parseInt(context.pathParam("accountId"));
-
-        boolean successful = accountService.deleteAccount(accountId);
+        Integer clientId = Integer.parseInt(context.pathParam("id"));
+        Client client = clientService.getClient(clientId);
+        Account account = accountService.getAccount(accountId);
+        boolean successful = false;
+        if((client != null && account != null) && client.getId() == account.getClientId()) {
+            successful = accountService.deleteAccount(accountId);
+        }
         if(successful) {
             context.contentType("application/json");
             context.result(ClientController.convertToJson(successful));
         }else {
             context.status(404);
+            context.result("Account or client not found.");
         }
     }
 
@@ -111,20 +149,21 @@ public class AccountController {
             account = accountService.withdrawFromBalance(accountId, withdrawAmount);
         } catch (Exception e)
         {
-            withdrawAmount = -1.0;
+            //withdrawAmount = -1.0;
         }
         try{
             depositAmount = Double.parseDouble(context.formParam("deposit"));
             account = accountService.depositToBalance(accountId, depositAmount);
         } catch(Exception e)
         {
-            depositAmount = -1.0;
+            //depositAmount = -1.0;
         }
 
         if(account == null){
             context.status(404);
         }else if(account.getId() == 0){
             context.status(422);
+            log.error("Insufficient funds in account.");
         }else{
             context.contentType("application/json");
             context.result(ClientController.convertToJson(account));
@@ -145,6 +184,92 @@ public class AccountController {
             context.contentType("application/json");
             context.result(ClientController.convertToJson(account));
         }
+    }
+
+    public static void updatePartialAccount(Context context) {
+        Integer clientId = 0;
+        Integer accountId = 0;
+        String name = "";
+        Double balance = 0.0;
+        Category category = Category.NONE;
+        Double withdrawAmount = 0.0;
+        Double depositAmount;
+        Account account = null;
+        Account updatedAccount = null;
+
+        try{
+            clientId = Integer.parseInt(context.pathParam("id"));
+            accountId = Integer.parseInt(context.pathParam("accountId"));
+        }catch (Exception e){
+            clientId = 0;
+            accountId = 0;
+        }
+
+        try{
+            withdrawAmount = Double.parseDouble(context.formParam("withdraw"));
+            account = accountService.withdrawFromBalance(accountId, withdrawAmount);
+        } catch (Exception e)
+        {
+            withdrawAmount = -1.0;
+        }
+        try{
+            depositAmount = Double.parseDouble(context.formParam("deposit"));
+            account = accountService.depositToBalance(accountId, depositAmount);
+        } catch(Exception e)
+        {
+            depositAmount = -1.0;
+        }
+
+        /*if(withdrawAmount < 0.00){
+            context.status(422);
+            context.result("No bueno muchacho.");
+            return;
+        }*/
+
+        if(account == null){
+            context.status(404);
+        }else if(account.getId() == 0){
+            context.status(422);
+        }else{
+            context.contentType("application/json");
+            context.result(ClientController.convertToJson(account));
+        }
+
+
+        try{
+            name = context.formParam("name");
+
+        }catch (Exception e){
+            name = "";
+        }
+        try{
+            balance = Double.parseDouble(context.formParam("balance"));
+
+        }catch (Exception e){
+            balance = -1.0;
+        }
+        try{
+            category = Category.valueOf(context.formParam("category").toUpperCase());
+
+        }catch (Exception e){
+            category = Category.NONE;
+        }
+
+        Account accountToUpdate = new Account(accountId, name, balance, category, clientId);
+        //if any fields need updated, update them.
+        if(accountToUpdate.getName() != null || accountToUpdate.getBalance() != -1.00 ||
+                !accountToUpdate.getCategory().equals(Category.NONE)){
+            updatedAccount = accountService.updateAccountPartial(accountId, name, balance, category, clientId);
+
+
+            if(updatedAccount == null){
+                context.status(404);
+            }else {
+                context.contentType("application/json");
+                context.result(ClientController.convertToJson(updatedAccount));
+            }
+        }
+
     }
 }
